backend/requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
sqlalchemy==2.0.19
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-cloud-aiplatform>=1.38.0
google-cloud-texttospeech==2.16.2
twilio==9.6.3
pytest==7.4.0
httpx==0.24.1
pytest-asyncio==0.21.1
aiosqlite==0.19.0
alembic==1.12.0
bcrypt==4.0.1
greenlet==3.0.3 